Refactor azureProfileData in store
