--- src/client/src/store/azureProfileData/azure/model.ts
@@ -1,6 +1,0 @@
-import { AZURE_TYPEKEYS } from "../typeKeys";
-
-export interface IAzureValidationStatusAction {
-  type: AZURE_TYPEKEYS.SET_VALIDATION_STATUS;
-  payload: boolean;
-}
