--- src/client/src/store/azureProfileData/azureActionType.ts
@@ -1,9 +1,0 @@
-import { ILoginToAzureAction, ILogoutAction } from "./login/model";
-import { IAzureValidationStatusAction } from "./azure/model";
-
-type AzureActionType =
-  | ILoginToAzureAction
-  | ILogoutAction
-  | IAzureValidationStatusAction;
-
-export default AzureActionType;
