--- src/client/src/store/azureProfileData/typeKeys.ts
@@ -1,7 +1,0 @@
-export enum AZURE_TYPEKEYS {
-  IS_LOGGED_IN_TO_AZURE = "WTS/azure/IS_LOGGED_IN_TO_AZURE",
-  LOG_OUT_OF_AZURE = "WTS/azure/LOG_OUT_OF_AZURE",
-  LOG_IN_TO_AZURE = "WTS/azure/LOG_IN_TO_AZURE",
-  GET_SUBSCRIPTION_DATA = "WTS/azure/GET_SUBSCRIPTION_DATA",
-  SET_VALIDATION_STATUS = "WTS/azure/SET_VALIDATION_STATUS"
-}
