+++ src/client/src/store/config/azure/azureActionType.ts
@@ -1,0 +1,7 @@
+import { ILoginToAzureAction, ILogoutAction } from "./model";
+
+type AzureActionType =
+  | ILoginToAzureAction
+  | ILogoutAction
+
+export default AzureActionType;
