+++ src/client/src/store/config/azure/typeKeys.ts
@@ -1,0 +1,4 @@
+export enum AZURE_TYPEKEYS {
+  LOG_OUT_OF_AZURE = "WTS/azure/LOG_OUT_OF_AZURE",
+  LOG_IN_TO_AZURE = "WTS/azure/LOG_IN_TO_AZURE"
+}
