+++ src/client/src/store/config/combineReducers.ts
@@ -2,7 +2,6 @@ import { combineReducers } from "redux";
 import detailsPage from "../navigation/routes/detailsPageReducer";
 import previewStatus from "./config/previewReducer";
 import validations from "./validations/setValidationsReducer";
-import isValidatingName from "./validations/validatingNameReducer";
 import versions from "./versions/reducer";
 import { azureProfileData } from "./azure/reducer";
 import appType from "./app/reducer";
@@ -11,7 +10,6 @@ const configReducer = combineReducers({
   detailsPage,
   previewStatus,
   validations,
-  isValidatingName,
   versions,
   azureProfileData,
   appType
